requirements.txt
# Linux 的 libxcb / libegl 等屬於系統層 runtime，需用套件管理器安裝，不在 pip 清單中替代這些 Python 套件。
PyQt5==5.15.11
PyQtWebEngine==5.15.7
opencv-python>=4.10
requests==2.33.1
websocket-client==1.9.0
websockets==16.0
cryptography==46.0.7
langchain==0.3.27
langchain-community==0.3.31
python-dotenv==1.2.2
azure-cognitiveservices-speech==1.46.0
